requirements.txt
Flask
Flask-Sockets
Flask-SQLAlchemy
Flask-Argon2
gevent
gunicorn
pycorenlp
